+++ src/vumeter_widget.c
@@ -119,9 +119,11 @@ static bool vumeter_select(widget_t *wdgt, int indx) {
             error_printf("failed to load VUMeter media, retrying in 1 second!");
             sleep_milli_seconds(1000);
             if (!vu_meters_load_media(wdgt->view->app->renderer, vw->meters[indx].vss)) {
-                unsigned texture_bytes = tcache_get_texture_bytes_count();
-                unsigned surface_bytes = tcache_get_surface_bytes_count();
-                error_printf("failed to load VUMeter media: texture cache memory: texture:%u %fMiB surface:%u %fMib\n", texture_bytes, (float)texture_bytes/(1024*1024), surface_bytes, (float)surface_bytes/(1024*1024));
+                size_t texture_bytes = tcache_get_texture_bytes_count();
+                size_t surface_bytes = tcache_get_surface_bytes_count();
+                error_printf("failed to load VUMeter media: texture cache memory: texture:%u %fMiB surface:%u %fMib\n",
+                        texture_bytes, (float)texture_bytes/(1024*1024),
+                        (unsigned)surface_bytes, (float)surface_bytes/(1024*1024));
             }
 //            exit(EXIT_FAILURE);
         }
